requirements.txt
customtkinter
obs-websocket-py
keyboard